Move emulator lifecycle logic to module.
